fix: restore FinCat and ACSet migration helpers used by DataMigrations
Restore the Catlab helpers that DataMigrations relies on under Catlab 0.17, including pointed schema handling, free-graph mapping, ACSet functor materialization, and Yoneda parsing updates.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>

+++ src/categorical_algebra/cats/categories/CoproductCats.jl
@@ -76,7 +76,7 @@ end
 
 GATlab.getvalue(x::NamedCoproductCat) = x.cats
 
-Base.getindex(x::NamedCoproductCat, i::Int) = x.cats[i]
+Base.getindex(x::NamedCoproductCat, i) = x.cats[i]
 
 @instance ThCategoryExplicitSets{O,H} [model::NamedCoproductCat{O,H}
                                              ] where {O,H} begin
